Convert HomePage to Typescript, Remove idioms state from HomePage and into TableSection
